requirements.txt
obsei[source,sink]
google-cloud-storage==1.38.0
